Always use "same" padding when computing spectral norm for conv layer.
This will be done no matter what padding the underlying convolutional layer
uses. This is because the choice of padding does not affect the spectral norm
of the convolutional kernel, but using "valid" padding does not preserve the
tensor sizes when going between `u` and `v` in the power iteration method.

PiperOrigin-RevId: 328230441